requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohttp==3.8.5
    # via
    #   kopf
    #   kubernetes-asyncio
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.7.22
    # via kubernetes-asyncio
charset-normalizer==3.2.0
    # via aiohttp
click==8.1.6
    # via kopf
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
iso8601==2.0.0
    # via kopf
kopf==1.36.2
    # via dask-operator (pyproject.toml)
kubernetes-asyncio==24.2.3
    # via dask-operator (pyproject.toml)
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
python-dateutil==2.8.2
    # via kubernetes-asyncio
python-json-logger==2.0.7
    # via kopf
pyyaml==6.0.1
    # via
    #   kopf
    #   kubernetes-asyncio
six==1.16.0
    # via
    #   kubernetes-asyncio
    #   python-dateutil
typing-extensions==4.7.1
    # via kopf
urllib3==2.0.4
    # via kubernetes-asyncio
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
